Deprecate flowWith with ERROR
